# docs/swagger.yaml
definitions:
...
  models.Post:
    properties:
      hashtags:
        items:
          type: string
        type: array
      is_blocked:
        type: boolean
      likes:
        type: integer
      media_uri:
        items:
          type: string
        type: array
      pid:
        type: string
      text:
        type: string
      timestamp:
        type: string
      uid:
        type: string
    type: object
  models.PostCreate:
    properties:
      hashtags:
        items:
          type: string
        type: array
      is_private:
        type: boolean
      media_uri:
        items:
          type: string
        type: array
      mentioned_user_ids:
        description: Agregar esta línea
        items:
          type: string
        type: array
      text:
        type: string
    type: object
  models.PostUpdate:
    properties:
      hashtags:
        items:
          type: string
        type: array
      is_private:
        type: boolean
      media_uri:
        items:
          type: string
        type: array
      text:
        type: string
    type: object
...
info:
  contact: {}
  title: SnapMsg API
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: multi
        description: hashtags
        in: query
        items:
          type: string
        name: hashtags
        type: array
      - description: author's nickname
        in: query
        name: nick
        type: string
      - description: text to match
        in: query
        name: text
        type: string
      - default: 100
        description: limit
        in: query
        maximum: 100
        minimum: 0
        name: limit
        required: true
        type: integer
      - default: 0
        description: page
        in: query
        minimum: 0
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      security:
      - Bearer: []
      summary: Get posts filtering by query
      tags:
      - posts methods
    post:
      consumes:
      - application/json
      parameters:
      - description: data for the new post
        in: body
        name: PostCreate
        required: true
        schema:
          $ref: '#/definitions/models.PostCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create a new post
      tags:
      - posts methods
  /posts/{pid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: post id to delete
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete post owned by current user
      tags:
      - posts methods
    patch:
      consumes:
      - application/json
      parameters:
      - description: post id to update
        in: path
        name: pid
        required: true
        type: string
      - description: data for update the post
        in: body
        name: PostUpdate
        required: true
        schema:
          $ref: '#/definitions/models.PostUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
      security:
      - Bearer: []
      summary: Update some fields of a post
      tags:
      - posts methods
  /posts/{pid}/snapshares:
    delete:
      consumes:
      - application/json
      parameters:
      - description: snapshare id to delete
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete snapshare owned by current user
      tags:
      - posts methods
  /posts/favs:
    get:
      consumes:
      - application/json
      parameters:
      - default: 100
        description: limit
        in: query
        maximum: 100
        minimum: 0
        name: limit
        required: true
        type: integer
      - default: 0
        description: page
        in: query
        minimum: 0
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      security:
      - Bearer: []
      summary: List user fav posts
      tags:
      - posts methods
  /posts/favs/{pid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: post id to unfav
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Remove a post from user favs
      tags:
      - posts methods
    get:
      consumes:
      - application/json
      parameters:
      - description: pid to check fav
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      security:
      - Bearer: []
      summary: Check if current user favs a given post
      tags:
      - posts methods
    post:
      consumes:
      - application/json
      parameters:
      - description: post id to mark as fav
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Add a post to user favs
      tags:
      - posts methods
  /posts/feed:
    get:
      consumes:
      - application/json
      parameters:
      - default: 100
        description: limit
        in: query
        maximum: 100
        minimum: 0
        name: limit
        required: true
        type: integer
      - default: 0
        description: page
        in: query
        minimum: 0
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      security:
      - Bearer: []
      summary: Get feed of the user making the request
      tags:
      - posts methods
  /posts/likes/{pid}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: post id to unlike
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Unlike a post
      tags:
      - posts methods
    get:
      consumes:
      - application/json
      parameters:
      - description: pid to check like
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      security:
      - Bearer: []
      summary: Check if current user liked a given post
      tags:
      - posts methods
    post:
      consumes:
      - application/json
      parameters:
      - description: post id to like
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Add a like to a post
      tags:
      - posts methods
  /posts/me:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: multi
        description: hashtags
        in: query
        items:
          type: string
        name: hashtags
        type: array
      - description: text to match
        in: query
        name: text
        type: string
      - default: 100
        description: limit
        in: query
        maximum: 100
        minimum: 0
        name: limit
        required: true
        type: integer
      - default: 0
        description: page
        in: query
        minimum: 0
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      security:
      - Bearer: []
      summary: Get posts filtering by query
      tags:
      - posts methods
  /posts/me/snapshares:
    get:
      consumes:
      - application/json
      description: Retrieves the snapshares associated with the current authenticated
        user.
      parameters:
      - default: 100
        description: Limit of snapshares to retrieve
        in: query
        maximum: 100
        minimum: 0
        name: limit
        type: integer
      - default: 0
        description: Page number for pagination
        in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get snapshares of current user
      tags:
      - posts methods
  /posts/me/stats:
    get:
      consumes:
      - application/json
      description: Get statistics for the current user's posts within a date range
      parameters:
      - description: Start Date
        format: date
        in: query
        name: start
        type: string
      - description: End Date
        format: date
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get current user's posts statistics
      tags:
      - stats methods
...
  /posts/snapshares/{pid}:
    get:
      consumes:
      - application/json
      description: Checks if the current authenticated user has snapshared a specific
        post.
      parameters:
      - description: Post ID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      security:
      - Bearer: []
      summary: Check if a post is snapshared by current user
      tags:
      - posts methods
    post:
      consumes:
      - application/json
      description: Creates a new snapshare for a given post by the current user.
      parameters:
      - description: Post ID
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create a snapshare
      tags:
      - posts methods
...
  /trendings:
    get:
      consumes:
      - application/json
      description: Retrieves a list of trending topics.
      parameters:
      - default: 10
        description: Limit of topics to retrieve
        in: query
        maximum: 100
        minimum: 0
        name: limit
        required: true
        type: integer
      - default: 0
        description: Page number for pagination
        in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      security:
      - Bearer: []
      summary: Get Trending Topics
      tags:
      - posts methods
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and Firebase JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"